moved resolver.js under lib
